Move datatypes package to root
