+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/datatypes/ItemStackPersistentDataType.kt
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.core.persistence.datatypes
+package io.github.pylonmc.pylon.core.datatypes
 
 import org.bukkit.inventory.ItemStack
 import org.bukkit.persistence.PersistentDataAdapterContext
