+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/datatypes/NamespacedKeyPersistentDataType.kt
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.core.persistence.datatypes
+package io.github.pylonmc.pylon.core.datatypes
 
 import org.bukkit.NamespacedKey
 import org.bukkit.persistence.PersistentDataAdapterContext
