+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/datatypes/VectorPersistentDataType.kt
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.core.persistence.datatypes
+package io.github.pylonmc.pylon.core.datatypes
 
 import io.github.pylonmc.pylon.core.pluginInstance
 import org.bukkit.NamespacedKey
