+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/serializer/SerializerTestBlockPosition.java
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.test.test.serializer;
 
-import io.github.pylonmc.pylon.core.persistence.datatypes.PylonSerializers;
+import io.github.pylonmc.pylon.core.datatypes.PylonSerializers;
 import io.github.pylonmc.pylon.core.util.position.BlockPosition;
 import io.github.pylonmc.pylon.test.PylonTest;
