+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/serializer/SerializerTestNamespacedKey.java
@@ -1,7 +1,7 @@
 package io.github.pylonmc.pylon.test.test.serializer;
 
 
-import io.github.pylonmc.pylon.core.persistence.datatypes.PylonSerializers;
+import io.github.pylonmc.pylon.core.datatypes.PylonSerializers;
 import io.github.pylonmc.pylon.test.PylonTest;
 import org.bukkit.NamespacedKey;
